requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
websockets
sentence-transformers
faiss-cpu
transformers
torch
numpy
scikit-learn
pandas
tqdm
regex
PyMuPDF
streamlit
ray
dask[complete]
pyspark
pytest
